fixed bug in SlideScene::identifier affecting scene distribution
SliderScene::identifier returned a wrong type. This could lead to the
creation of new scenes when an existing scene should be used instead.

+++ src/slidescene.h
@@ -167,7 +167,7 @@ public:
     {return master;}
 
     /// Hash for this scene based on shift, master and page_part.
-    unsigned int identifier() const
+    size_t identifier() const
     {return qHash(QPair<int, const void*>(shift, master)) + page_part;}
 
     /// Currently visible page.
